Parse response from simple endpoint as HTML
Initially we parsed the response as XML. I thought it'd be sufficient
as the specs says that it's a simple page with only anchor tags.
However, HTML attributes can contain unescaped > and < symbols. This is
not compliant XML and therefore parsing fails. (those symbols are used
when specifying required python version)